refactor(inbox): move the health record into its own module
`src/inbox.rs` had reached 1934 lines (808 implementation, 1126 tests), well
past the 800-line guideline the repo follows elsewhere, and it held two
genuinely separate concerns.

`src/inbox/health.rs` now holds `InboxHealth`, `BlindWindow` and the per-relay
pacing — the half `scheduler.rs` actually depends on. `src/inbox/mod.rs` keeps
the subscription, the keeper and the watchdog audit. The module is private and
re-exports `InboxHealth`, `InboxStatus` and `InstallError`, so `crate::inbox::`
paths are unchanged for callers.

Behaviour is identical: the test helper that reached into the backoff map is
now `InboxHealth::is_backing_off` under `cfg(test)`, and `at`, `observe`,
`note_relay_acknowledged_at` and `now_secs` widen from private to `pub(super)`
so the keeper's half can still reach them.
